pyproject.toml
[project]
name = "car-recog"
version = "0.0.1"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "matplotlib",
    "pillow",
]
authors = [
    { name="Iosiv42", email="VIZmuth@tuta.io" }
]
readme = "README.md"

[project.scripts]
car-recog = "car_recog:main_cli"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
